[silgen] Extract SILGenBuilder into its own header/cpp file.
NFC.
